progress

+++ src/notifier/db/tables.js
@@ -19,7 +19,7 @@ export const Tables = [
     )`,
     `CREATE TABLE IF NOT EXISTS users (
         user_id TEXT PRIMARY KEY,
-        discord_id INTEGER UNIQUE,
+        discord_id BIGINT UNIQUE,
         is_channel BOOLEAN NOT NULL
     )`,
     `CREATE TABLE IF NOT EXISTS notification_strategies (
